+++ HISTORY.rst
@@ -7,6 +7,7 @@ v0.2.0 (NEXT)
 
 * For continuous summary datasets, rename ``responses`` array to ``means``
 * Added quickstart for using R
+* Added the ability to enter continuous individual datasets
 
 
 v0.1.0 (2016-12-05)